backend/src/agents/multi_ai_gateway.py
"""
Systematic AI Agent Gateway for iGSIM
2026 AI Agent Standard Implementation
Brand: iGSIM AI Agent powered by eSIM Myanmar
"""
import os
from typing import Dict, Any
import google.generativeai as genai
import openai
import groq
from dataclasses import dataclass
from enum import Enum

class AIProvider(Enum):
    GEMINI = "gemini"
    OPENAI = "openai"
    GROQ = "groq"
    XAI = "xai"

@dataclass
class AIResponse:
    provider: AIProvider
    content: str
    tokens_used: int
    latency_ms: float

class iGSIMAIGateway:
    """Systematic AI Agent Gateway following 2026 standards"""
    
    def __init__(self):
        self._initialize_providers()
        self._setup_models()
        
    def _initialize_providers(self):
        # Google Gemini
        genai.configure(api_key=os.getenv('GEMINI_API_KEY'))
        
        # xAI
        self.xai_api_key = os.getenv('XAI_API_KEY')
        
        # Groq
        self.groq_client = groq.Groq(
            api_key=os.getenv('GROQ_API_KEY')
        )
        
        # OpenAI compatible
        openai.api_key = os.getenv('XAI_API_KEY')
        
    def _setup_models(self):
        self.models = {
            AIProvider.GEMINI: {
                'creative': 'gemini-pro',
                'technical': 'gemini-pro'
            },
            AIProvider.GROQ: {
                'fast': 'mixtral-8x7b-32768',
                'quality': 'llama2-70b-4096'
            }
        }
    
    async def process_esim_request(self, query: str, context: Dict[str, Any]) -> AIResponse:
        """Process eSIM M2M AI requests"""
        provider = self._select_provider(query, context)
        
        if provider == AIProvider.GEMINI:
            return await self._call_gemini(query, context)
        elif provider == AIProvider.GROQ:
            return await self._call_groq(query, context)
        
    async def _call_gemini(self, query: str, context: Dict[str, Any]) -> AIResponse:
        """Systematic Gemini API integration"""
        model = genai.GenerativeModel(self.models[AIProvider.GEMINI]['technical'])
        
        prompt = f"""
        iGSIM eSIM AI Agent Request:
        Context: {context}
        Query: {query}
        
        Provide systematic, accurate response for eSIM M2M services.
        """
        
        response = model.generate_content(prompt)
        return AIResponse(
            provider=AIProvider.GEMINI,
            content=response.text,
            tokens_used=response.usage_metadata.total_token_count,
            latency_ms=0
        )
