backend/requirements.txt
# Core Framework
PyQt6==6.5.0
PySide6==6.5.0
fastapi==0.104.1
uvicorn==0.24.0

# AI and ML
google-generativeai==0.3.0
openai==1.3.0
groq==0.3.0

# Google Cloud
google-cloud-aiplatform==1.38.0
google-cloud-firestore==2.13.0
google-cloud-functions==1.13.0
firebase-admin==6.2.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
aiohttp==3.9.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
